+++ packages/components/src/elements/RadioGroup/RadioGroup.ts
@@ -3,7 +3,7 @@ import { customElement, property } from "lit/decorators.js";
 import { StrataElement } from "../StrataBase.js";
 
 @customElement("strata-radio-group")
-export class RadioGroup extends StrataElement {
+export class StrataRadioGroup extends StrataElement {
   @property({type:String}) label="";
   render(){return html`<fieldset><legend>${this.label}</legend><slot></slot></fieldset>`;}
 }